pyproject.toml
[project]
name = "sbsys-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "SimplyPerry", email = "andr2999@hotmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "python-dotenv>=1.1.0",
    "httpx>=0.27.0",
    "authlib>=1.4.0"
]

[dependency-groups]
dev = [
    "pytest>=8.3.4",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
